testing

+++ libft/ft_putnbr_base_u.c
@@ -6,7 +6,7 @@
 /*   By: bcanals- <bcanals-@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 15:59:46 by bcanals-          #+#    #+#             */
-/*   Updated: 2024/10/13 03:35:23 by bizcru           ###   ########.fr       */
+/*   Updated: 2024/10/13 04:48:27 by bizcru           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ int	check_arguments(char *base)
 	return (1);
 }
 
-int	ft_putnbr_base_easy(unsigned int nbr, char *base, unsigned int size, int *count)
+int	ft_putnbr_base_easy(size_t nbr, char *base, unsigned int size, int *count)
 {
 	char		c;
 
@@ -66,7 +66,7 @@ int	ft_putnbr_base_easy(unsigned int nbr, char *base, unsigned int size, int *co
 	return (++*count);
 }
 
-int	ft_putnbr_base_u(unsigned int nbr, char *base)
+int	ft_putnbr_base_u(size_t nbr, char *base)
 {
 	unsigned int	size;
 	char			c;
